requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
genanki==0.13.1
wordfreq==3.1.1
httpx==0.26.0
jinja2==3.1.3
python-dotenv==1.0.0
python-multipart==0.0.6
